// services/shared/types/common.ts
export interface LogEntry {
  timestamp: string;
  level: LogLevel;
  service: string;
  version: string;
  traceId?: string | undefined;
  spanId?: string | undefined;
  userId?: string | undefined;
  requestId: string;
  message: string;
  metadata?: Record<string, any>;
  error?: LogError;
}

/**
 * ログレベル
 */
export enum LogLevel {
  DEBUG = 'debug',
  INFO = 'info',
  WARN = 'warn',
  ERROR = 'error'
}

/**
 * ログエラー
 */
export interface LogError {
  name: string;
  message: string;
  stack: string;
}